# backend/api/cobytes-openapi.yml
openapi: 3.0.3
info:
  title: Cobytes Security Platform API
  description: |
    The Cobytes Security Platform API provides comprehensive security scanning and reporting capabilities.
    
    ## Authentication
    All endpoints (except health and API info) require JWT authentication.
    Use the `/api/auth/login` endpoint to obtain a token.
    
    ## Base URL
    - Development: `http://localhost:3001`
    - Production: `https://api.cobytes.com`
  version: 1.0.0
  contact:
    name: Cobytes API Support
    email: api-support@cobytes.com
    url: https://docs.cobytes.com
  license:
    name: Proprietary
    url: https://cobytes.com/license
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.cobytes.com
    description: Production server
...
paths:
...
  /api/reports:
    get:
      tags:
        - Reports
      summary: Get All Reports
      description: Get a list of all generated reports
      operationId: getAllReports
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/reports/generate:
    post:
      tags:
        - Reports
      summary: Generate Report
      description: Generate a new report from one or more scans
      operationId: generateReport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateReportRequest'
      responses:
        '202':
          description: Report generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/reports/templates/available:
    get:
      tags:
        - Reports
      summary: Get Available Templates
      description: Get a list of available report templates
      operationId: getReportTemplates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplatesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/reports/{reportId}:
    get:
      tags:
        - Reports
      summary: Get Report Details
      description: Get detailed information about a specific report
      operationId: getReportById
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          description: Report ID
          schema:
            type: string
      responses:
        '200':
          description: Report details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDetailsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Reports
      summary: Delete Report
      description: Delete a report
      operationId: deleteReport
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          description: Report ID
          schema:
            type: string
      responses:
        '200':
          description: Report deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/reports/{reportId}/download:
    get:
      tags:
        - Reports
      summary: Download Report
      description: Download a generated report
      operationId: downloadReport
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          description: Report ID
          schema:
            type: string
      responses:
        '200':
          description: Report content
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDownloadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/reports/stats/overview:
    get:
      tags:
        - Reports
      summary: Get Report Statistics
      description: Get statistics about all reports
      operationId: getReportStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Report statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Vulnerability:
      type: object
      properties:
        severity:
          type: string
          enum: [critical, high, medium, low, info]
          example: high
        title:
          type: string
          example: SQL Injection vulnerability
        description:
          type: string
          example: Potential SQL injection in login form
        cve:
          type: string
          example: CVE-2024-1234
        cwe:
          type: string
          example: CWE-89
        affectedAssets:
          type: array
          items:
            type: string
          example: [192.168.1.100, https://example.com/login]
        recommendation:
          type: string
          example: Use parameterized queries
...
    GenerateReportRequest:
      type: object
      required:
        - scanIds
      properties:
        scanIds:
          type: array
          items:
            type: string
          example: [scan_1748799588916_dw04msszk]
        format:
          type: string
          enum: [pdf, html, json, csv]
          default: pdf
          example: pdf
        includeDetails:
          type: boolean
          default: true
          example: true
        template:
          type: string
          example: executive_summary
    Report:
      type: object
      properties:
        reportId:
          type: string
          example: report_1748799588923_9rehl4b0k
        name:
          type: string
          example: Security Report 2025-06-01
        scanIds:
          type: array
          items:
            type: string
        format:
          type: string
          enum: [pdf, html, json, csv]
          example: pdf
        status:
          type: string
          enum: [generating, completed, failed]
          example: generating
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        size:
          type: integer
          example: 245760
        downloadUrl:
          type: string
          nullable: true
          example: /api/reports/report_1748799588923_9rehl4b0k/download
    ReportResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Report'
        message:
          type: string
          example: Report generation started
    ReportsListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Report'
        count:
          type: integer
          example: 1
    Template:
      type: object
      properties:
        id:
          type: string
          example: executive_summary
        name:
          type: string
          example: Executive Summary
        description:
          type: string
          example: High-level overview for executives
        formats:
          type: array
          items:
            type: string
            enum: [pdf, html, json, csv]
          example: [pdf, html]
    TemplatesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Template'
    ReportDetailsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Report'
    ReportContent:
      type: object
      properties:
        summary:
          type: object
          properties:
            totalVulnerabilities:
              type: integer
              example: 15
            criticalCount:
              type: integer
              example: 2
            highCount:
              type: integer
              example: 5
            mediumCount:
              type: integer
              example: 6
            lowCount:
              type: integer
              example: 2
            scanDate:
              type: string
              format: date-time
            targetCount:
              type: integer
              example: 1
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/Vulnerability'
        recommendations:
          type: array
          items:
            type: string
    ReportDownloadResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            reportId:
              type: string
              example: report_1748799588923_9rehl4b0k
            name:
              type: string
              example: Security Report 2025-06-01
            format:
              type: string
              example: pdf
            content:
              $ref: '#/components/schemas/ReportContent'
            generatedAt:
              type: string
              format: date-time
    ReportStatsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            totalReports:
              type: integer
              example: 1
            reportsByFormat:
              type: object
              properties:
                pdf:
                  type: integer
                  example: 1
                html:
                  type: integer
                  example: 0
                json:
                  type: integer
                  example: 0
                csv:
                  type: integer
                  example: 0
            reportsByStatus:
              type: object
              properties:
                generating:
                  type: integer
                  example: 0
                completed:
                  type: integer
                  example: 1
                failed:
                  type: integer
                  example: 0
            averageGenerationTime:
              type: number
              example: 3.2
            totalStorageUsed:
              type: integer
              example: 245760
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Resource deleted successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message
        code:
          type: string
          example: ERROR_CODE
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Validation failed
            code: VALIDATION_ERROR
            details:
              field: Error details
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Invalid or missing authentication
            code: UNAUTHORIZED
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Resource not found
            code: NOT_FOUND
...