# backend/src/routes/report.routes.ts
import { Router, Request, Response } from 'express';
import { logger } from '@utils/logger';
import { asyncHandler } from '@middleware/asyncHandler';

const router = Router();

// In-memory storage for reports
const reports = new Map<string, any>();

// Report interfaces
interface ReportRequest {
  scanIds: string[];
  format: 'pdf' | 'html' | 'json' | 'csv';
  includeDetails: boolean;
  template?: string;
}

// Get all reports
router.get('/', asyncHandler(async (req: Request, res: Response) => {
  logger.info('GET /api/reports - Fetching all reports');
  
  const reportsList = Array.from(reports.values()).map(report => ({
    reportId: report.reportId,
    name: report.name,
    format: report.format,
    createdAt: report.createdAt,
    status: report.status,
    size: report.size
  }));
  
  res.json({
    success: true,
    data: reportsList,
    count: reportsList.length
  });
}));

// Get report by ID
router.get('/:reportId', asyncHandler(async (req: Request, res: Response) => {
  const { reportId } = req.params;
  logger.info(`GET /api/reports/${reportId} - Fetching report`);
  
  const report = reports.get(reportId);
  
  if (!report) {
    return res.status(404).json({
      success: false,
      error: 'Report not found'
    });
  }
  
  res.json({
    success: true,
    data: report
  });
}));

// Generate new report
router.post('/generate', asyncHandler(async (req: Request, res: Response) => {
  const reportRequest: ReportRequest = req.body;
  logger.info('POST /api/reports/generate - Generating new report', { 
    format: reportRequest.format,
    scanCount: reportRequest.scanIds?.length 
  });
  
  // Validate request
  if (!reportRequest.scanIds || reportRequest.scanIds.length === 0) {
    return res.status(400).json({
      success: false,
      error: 'At least one scan ID is required'
    });
  }
  
  if (!reportRequest.format) {
    reportRequest.format = 'pdf';
  }
  
  // Create report
  const reportId = `report_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`;
  const newReport = {
    reportId,
    name: `Security Report ${new Date().toISOString().split('T')[0]}`,
    scanIds: reportRequest.scanIds,
    format: reportRequest.format,
    status: 'generating',
    createdAt: new Date(),
    updatedAt: new Date(),
    size: 0,
    downloadUrl: null
  };
  
  reports.set(reportId, newReport);
  
  // Simulate report generation
  setTimeout(() => {
    const report = reports.get(reportId);
    if (report) {
      report.status = 'completed';
      report.updatedAt = new Date();
      report.size = Math.floor(Math.random() * 1000000) + 100000; // Random size between 100KB and 1MB
      report.downloadUrl = `/api/reports/${reportId}/download`;
      
      // Add mock report content
      report.content = {
        summary: {
          totalVulnerabilities: 15,
          criticalCount: 2,
          highCount: 5,
          mediumCount: 6,
          lowCount: 2,
          scanDate: new Date(),
          targetCount: reportRequest.scanIds.length
        },
        vulnerabilities: [
          {
            severity: 'critical',
            title: 'Remote Code Execution',
            cve: 'CVE-2024-1234',
            description: 'Unauthenticated remote code execution vulnerability',
            affectedAssets: ['192.168.1.100', '192.168.1.101'],
            recommendation: 'Apply security patch immediately'
          },
          {
            severity: 'high',
            title: 'SQL Injection',
            cwe: 'CWE-89',
            description: 'SQL injection vulnerability in login form',
            affectedAssets: ['https://example.com/login'],
            recommendation: 'Use parameterized queries'
          }
        ],
        recommendations: [
          'Implement regular security patching schedule',
          'Enable Web Application Firewall (WAF)',
          'Conduct regular security assessments'
        ]
      };
    }
  }, 3000); // Generate after 3 seconds
  
  res.status(202).json({
    success: true,
    data: newReport,
    message: 'Report generation started'
  });
}));

// Download report
router.get('/:reportId/download', asyncHandler(async (req: Request, res: Response) => {
  const { reportId } = req.params;
  logger.info(`GET /api/reports/${reportId}/download - Downloading report`);
  
  const report = reports.get(reportId);
  
  if (!report) {
    return res.status(404).json({
      success: false,
      error: 'Report not found'
    });
  }
  
  if (report.status !== 'completed') {
    return res.status(400).json({
      success: false,
      error: 'Report is not ready for download',
      status: report.status
    });
  }
  
  // In a real app, this would serve the actual file
  // For demo, we'll return the report content as JSON
  res.json({
    success: true,
    data: {
      reportId: report.reportId,
      name: report.name,
      format: report.format,
      content: report.content,
      generatedAt: report.updatedAt
    }
  });
}));

// Get report templates
router.get('/templates/available', asyncHandler(async (req: Request, res: Response) => {
  logger.info('GET /api/reports/templates/available - Fetching report templates');
  
  const templates = [
    {
      id: 'executive_summary',
      name: 'Executive Summary',
      description: 'High-level overview for executives',
      formats: ['pdf', 'html']
    },
    {
      id: 'technical_detailed',
      name: 'Technical Detailed Report',
      description: 'Comprehensive technical analysis with remediation steps',
      formats: ['pdf', 'html', 'json']
    },
    {
      id: 'compliance_report',
      name: 'Compliance Report',
      description: 'Compliance-focused report (PCI-DSS, HIPAA, etc.)',
      formats: ['pdf', 'html']
    },
    {
      id: 'vulnerability_list',
      name: 'Vulnerability List',
      description: 'Simple list of vulnerabilities for tracking',
      formats: ['csv', 'json', 'pdf']
    }
  ];
  
  res.json({
    success: true,
    data: templates
  });
}));

// Delete report
router.delete('/:reportId', asyncHandler(async (req: Request, res: Response) => {
  const { reportId } = req.params;
  logger.info(`DELETE /api/reports/${reportId} - Deleting report`);
  
  if (!reports.has(reportId)) {
    return res.status(404).json({
      success: false,
      error: 'Report not found'
    });
  }
  
  reports.delete(reportId);
  
  res.json({
    success: true,
    message: 'Report deleted successfully'
  });
}));

// Get report statistics
router.get('/stats/overview', asyncHandler(async (req: Request, res: Response) => {
  logger.info('GET /api/reports/stats/overview - Fetching report statistics');
  
  const stats = {
    totalReports: reports.size,
    reportsByFormat: {
      pdf: Array.from(reports.values()).filter(r => r.format === 'pdf').length,
      html: Array.from(reports.values()).filter(r => r.format === 'html').length,
      json: Array.from(reports.values()).filter(r => r.format === 'json').length,
      csv: Array.from(reports.values()).filter(r => r.format === 'csv').length
    },
    reportsByStatus: {
      generating: Array.from(reports.values()).filter(r => r.status === 'generating').length,
      completed: Array.from(reports.values()).filter(r => r.status === 'completed').length,
      failed: Array.from(reports.values()).filter(r => r.status === 'failed').length
    },
    averageGenerationTime: 3.2, // seconds
    totalStorageUsed: Array.from(reports.values()).reduce((sum, r) => sum + (r.size || 0), 0)
  };
  
  res.json({
    success: true,
    data: stats
  });
}));

export const reportRoutes = router;